sm8350-common: Use the ?= operator so BT properties can be overridden
System properties set with the "=" operator cannot be overriden. This
device is used elsewhere in the system and some of the properties need
to be overridden. The "?=" operator will use the assigned value if no
other values are assigned elsewhere, allowing for overrides.

Bug: 230638462
Test: adb shell getprop | grep bluetooth
Merged-In: Ie29700e8b87f6f3fc2da741411dfbe94fa56ca36
Change-Id: Ie29700e8b87f6f3fc2da741411dfbe94fa56ca36